requirements.txt
Django==1.5.1
South==0.8.1
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-extra-views==0.6.3
django-mathfilters==0.1.3
django-social-auth==0.7.23
gunicorn==17.5
hashlib==20081119
httplib2==0.8
oauth2==1.5.211
psycopg2==2.5.1
python-openid==2.2.5
static==0.4
wsgiref==0.1.2
